Minor bug-fixes: 1. double vs. size_t mix-up. 2. Commenting error.

+++ src/basic/source/cell.cpp
@@ -697,7 +697,7 @@ void Cell::doCollision(Particle *p, point_t& r, point_t& v, const point_t &force
     cout << "Cell::doCollision: tracking particle " << TRACK_PARTICLE << " !!! " << m_all_walls.size() << " walls in cell." << endl;
 
 #endif
-  double iterations = 0;
+  size_t iterations = 0;
   while (hit) {      
     ++iterations;
     if(iterations > 100)
